cleanup: renamed UserProfile component(s)
